requirements.txt
click==8.0.0
dnspython==2.1.0
Flask==1.1.0
Flask-PyMongo==2.3.0
itsdangerous==2.0.0
Jinja2==3.0.0
MarkupSafe==2.0.0
pymongo==3.11.4
python-dotenv==0.17.1
Werkzeug==2.0.0
bcrypt==3.2.0
Flask-RESTful==0.3.8
PyJWT==1.7.1
requests==2.25.1